ensure we re-print the candidate in case it didn't change
When printing a candidate, we do some optimizations already, such as:

- `bg-[var(--foo)]` -> `bg-(--foo)`
- `bg-[rgb(0,_0,_0)]` -> `bg-[rgb(0,0,0)]`

Consistency in your project will reduce the file size.

We parse and reprint the candidate if nothing changed during migrations
because we don't have dedicated migrations for them.

So (a lot) of these classes weren't fully updated to the v4 flavor of
the classes.

Luckily re-parsing the candidate is fast because we are re-using the
design system which means that we have a cached version of the
candidate.