cache candidate migrations
This is a first step in improving the performance of the upgrade tool.
Essentially let's cache the incoming candidate to the outgoing migration
based on the given inputs.

Notes: this is technically going to be slower now, because the
`location` will be different for every candidate right now.

In the next commits we will make sure to drop the `location` from the
cache (increasing the cache hits) and checking whether the migration is
safe ahead of time.