check safe migrations first
Instead of checking whether a candidate migration is safe during each
migration step, we can do it all up front. To make things worse, we did
check the safety in some migrations, but not in others which means that
we could still be performing unsafe migrations.

The big issue with "unsafe" migrations is where we use very small
classes such as `blur` or `visible` and they can be used in non-Tailwind
CSS contexts. Migrating these will result in unsafe migrations.

The moment we use variants, modifiers, arbitrary values the chances that
these are actual unsafe migrations are so low. We can use that as a
heuristic to check whether a migration is safe or not.

When we are dealing with `blur`, `visible` or `flex` we can apply the
other, existing, safety checks.

Once we know that a migration is unsafe, we will skip the migration.

Bonus points: this also takes the `location` information out of the
cache, which means that the cache will be smaller and increase cache
hits tremendously.