verify that the migrated class actually exists
When you have something like:
```
if (!duration) {}
```

It could be that we migrate this to:
```
if (duration!) {}
```

Hopefully our safety checks do not do that though. But in the event that
they do, the `!duration` should not be converted to `duration!` because
that class simply doesn't exist.

It wil parse correctly as:
```
[
  {
    "kind": "functional",
    "root": "duration",
    "modifier": null,
    "value": null,
    "variants": [],
    "important": true,
    "raw": "!duration"
  }
]
```

And that's because the `duration-<number>` _does_ work and produces
output. But this on its won won't produce any output at all.

If that's the case, then we can throw away any migrations related to
this and return the original value.