improve safe migration check
When a candidate doesn't parse, we throw it away. However, if you are in
a Tailwind CSS v3 project then it could be that we are dealing with
classes that don't exist in Tailwind CSS v4 and will be migrated later.

It could also be that we are dealing with legacy syntax, e.g.:
`tw__flex` if you have a custom variant separator.

This fixes that by only bailing in Tailwind CSS v4 and up.