ensure `group` and `peer` are valid during v3 migration
In order to properly parse the value, we already had to make sure that
`group` and `peer` exist in the framework during a Tailwind CSS v3
→ Tailwind CSS v4 migration.

We now also verify that the actual produced class is valid, for this we
use `@apply` internally, but because of the `return null` this was
considered invalid.

This fixes that by making sure that we have some declarations, in this
case I used a random key like `--phantom-class: group`.

Then to make sure that `group` and `group/foo` are considered distinct
classes, I also added the modifier to the equation.