allow upgrades in `pnpm` workspace root
When using pnpm with a workspace, if you happen to run the upgrade tool
in the root because some Tailwind CSS dependencies live there, then you
will get an error that you need the `--workspace-root` flag.

However, if we _always_ add it, then the upgrade tool will add
dependencies to the root, regardless of whether you are in a
sub-packages.

With this, the check doesn't happen and no error is thrown. This will
just install/bump the dependencies where necessary.