do not assume `tail` is available
Right now if a candidate looks like `[--btn-border:var(--color-maroon)/90)]` then there is a bad closing `)`.

In this case the value parser is parsing `var(--color-maroon)/90)`, and
because of the `)` we will hit the "closing function" branch. This
branch currently assumes that an opening `(` was used.

I am actually not 100% sure what we should do here because I don't think
there is a good solution. We could throw an error with more info, but
since this kind of class could be returned from Oxide you may not have
control over this at all.