CSS functions in fallback values are allowed
Before this, the moment we saw a `(` we stopped trying to convert
`bg-[--foo()]` to `bg-(--foo())`. This is because it doesn't make sense
to wrap a CSS function in another `var(…)`.

However, if that `(` exists in the fallback value, then it's fine and
we can still perform the conversion:

- `bg-[--my-color,theme(colors.red.500)]` → `bg-(--my-color,theme(colors.red.500))`