fix watchers never being cleaned up
Previously, there was an issue, where a slow compilation could cause the
watchers to build up and never be cleaned. This would result in both
excessive logging and the compilation time increasing exponentially.

This fix is a naive patch, simply locking if there is a build in
progress, and unlocking when it completes.