refactor: when migrating theme key, handle special cases first
We were handling special cases (keyframes, container) _after_ we handled
the normal theme keys. When handling the normal theme keys we skipped
the special keys which means that you have to keep the list in sync.

Now we will handle special cases first.

Another refactor here is that we will push data to an intermediate
variable before emitting the `@theme {…}` wrapper. This will allow us to
only emit CSS when needed.

We were explicitly returning `null` in some cases, but we can also just
return the empty string when there is no CSS to get the same effect.